feat: hook discussions provider list up to API
Prior to this, the list of providers in the UI was hardcoded in the api.js layer.

This commit hooks that up to our actual API endpoint, allowing us to load the provider list from the server.

It also - out of necessity - changes the way the AppCards are displayed, and what content is in them.  This was somewhat opportunistic as our design for them simplified anyway, no longer requiring a logo or a few of the other fields.

Because the actual API sends us less display-oriented data (i.e., no names or descriptions for things), we had to modify FeaturesTable and AppCard to fetch these strings from the messages file based on the app and feature IDs.

I’m not super thrilled with this approach, since it’s somewhat brittle.  Unexpected providers won’t display properly.  In the long run, I expect all this may come from some other system.  Using translations to load dynamic strings isn’t quite what it was intended for - i.e., we’re not putting “Piazza” in there because we want to translate it, but because the backend didn’t tell us what to use for the key “piazza”.