fix: change step two label to “Settings”
The mockups changed the label, which is great, since we’re aiming to remove the need for selectedAppId to be in DiscussionsSettings.  This is one step toward that.

This has the side effect of making app.name no longer necessary, which is great, since the app no longer contains its own name string.