refactor: add a context to provide a formRef to AppConfigForm components
This is an intermediary commit - it’ll likely fail linting since the whitespace is a bit off.  It’s also not doing anything yet - a subsequent commit will make use of this code (the provider and new ApplyButton component)

Today the DiscussionsSettings component needs to be responsible for form submission because it owns the “Apply” button that a user clicks to submit the form.  It declares a “formRef” which is passed into the form and attached to the form DOM element, and then the “Apply” button can call formRef.current.submit() to submit the form from DiscussionsSettings, even though the form itself is declared way down in a sub-module.

This is awkward, and forces this top level component to be smarter than it should be, and to know too much about its children.

An alternate pattern is to create a context provider (AppConfigFormProvider) which owns the formRef and is part of the sub-module.  Then create an AppConfigFormApplyButton component that knows how to pull that formRef out of the context and call submit on it.

This way, DiscussionsSettings doesn’t know anything about form submission, and we can move all that code into the app-config-form sub-module that should know about it.  Win!

Again, this commit just adds the code and inserts the provider into DiscussionsSettings - it doesn’t use it all yet.