fix: use the correct formRef for AppConfigForm children
Straight up refactoring bug.  This should be using the formRef from the context, but was still defining its own and passing it down.  This ultimately meant that the submit button wasn’t properly hooked up to the form, since the form was given a different ref than the submit button.