feat: handle access denied by setting DENIED statuses
This takes us one step closer to user messaging for permission denied errors by setting an explicit DENIED status and saveStatus when a request was denied because the user didn’t have permissions.  Note that this is different than a 401, which is unauthorized, meaning the user is logged out.  This doesn’t handle 401s.

Following this, we can then use the DENIED status to give the user feedback on what’s going on.