feat: adding error alerts
This commit adds friendly error alerts for connection and permission denied errors in the discussions app.

If the initial fetch apps request errors out or is denied, then the entire contents of the modal is replaced with an error message.

If the save app config request errors out, then a message is displayed at the top of the form.

If the sae app config request is denied, the entire contents of the modal is replaced with a permission denied error message, as in the first case.