fix: update discussions API handler to match shape from server (#109)
The application was expecting a different API shape than the server was sending.  I have _no_ idea how this made it through, as it’s blatantly broken.  I don’t see any changes in this code in the past few weeks.

Anyway, the client was expecting a “features” array inside an app (provider), which simply didn’t exist.  The app itself is the array.  This PR udpates the shape in `normalizeApps` and also the associated test, which was also equally as wrong.  Somehow.