refactor: update .env file with falsy values
More detail - if a variable isn’t defined in internal, these values will get used as fallbacks.  They’re also bad fallbacks, unfortunately.  When we originally set this system up, we didn’t realize that null would actually be cast to "null" when it gets into the MFE… meaning the string “null”, which is _truthy_, and very much not the same as null.  It’s a good practice to replace these `null`’s with `''` (an empty string) which will actually evaluate to be falsy.

Note that they only ever get used as fallbacks if we’ve screwed something up, so the existence of this file is kind of misleading - we always want the values from internal to be used.