fix: avoid onresize height loop
Open edX sets a global `min-height: 100%` for the body element (a common
strategy to allow for background images to fill the viewport), but this
is undesirable for iframes where the `window.onresize` is used to
programmatically adjust the iframe height.  Because `onresize` will fire when the body
adjusts itself as per its `min-height: 100%`, in certain situations this
will lead to an infinite resizing loop.  This is notably visible when:

* Rendering the video block preview when using Firefox: this causes the iframe
  size to jiggle indefinitely.
* Rendering the HTML block preview when there are empty block elements (such as
  `<p></p>`) that have a vertical margin set (which happens to be the
  default to HTML blocks): this will cause the iframe to increase in
  height infinitely.

By adding an inline `min-height: auto` on the wrapping body element,
this overrides the Studio-served CSS, and thus fixes the problem.