chore: Upgrade all dependencies
This upgrades all dependencies to their latest versions (notably Paragon
19.9 and React 17) in preparation for the move to Node 16 and dropping
IE 11 support.

As a result, several fixes to the codebase were required.  Including:

* Fixing Paragon components that changed;
* Removing the old Paragon-provided edX theme (breaks styling on
  many non-Paragon components, and will require revisiting);
* Fixes resulting from a new and stricter PropTypes;
* Adding a new configuration variable required by
  `frontend-component-footer`;
* Using the new `initializeMockApp` when testing;
* Moving from `babel-polyfill` to `corejs` (not strictly necessary, but
  this seemed like an opportune moment to do so);

Of note, `package-lock.json` was removed and regenerated using Node 12.

And finally, the list of known issues in the README was updated to
include ones caused by, or just found during, this dependency upgrade.